functions/worker/requirements.txt
functions-framework>=3.0.0
pandas>=2.0.0
requests>=2.31.0
google-cloud-bigquery>=3.11.0
pyarrow>=12.0.0
